requirements.txt
# The order of packages is significant, because pip processes them in the order
# of appearance. Changing the order has an impact on the overall integration
# process, which may cause wedges in the gate later.
beautifulsoup4>=4.6.0 # MIT
lxml!=3.7.0,>=4.1.1 # BSD
oslo.log>=3.37.0 # Apache-2.0
requests>=2.18.4 # Apache-2.0
PyYAML>=3.12 # MIT
